backend/requirements-core.txt
fastapi
uvicorn
pydantic==2.*
python-dotenv
bcrypt
python-jose[cryptography]
python-multipart
httpx
PyYAML
requests
email-validator==2.1.0.post1
Pillow
google-genai
google-cloud-aiplatform
openai
bs4
grpcio>=1.51.0